+++ gradle.properties
@@ -1,7 +1,7 @@
 # properties used in build.gradle
 
 # application properties
-mainClass=edu.oregonstate.mist.webapiskeleton.SkeletonApplication
+mainClass=edu.oregonstate.mist.slashcourses.SkeletonApplication
 archivesBaseName=web-api-skeleton
 
 # library properties
