+++ src/main/groovy/edu/oregonstate/mist/slashcourses/core/Sample.groovy
@@ -1,4 +1,4 @@
-package edu.oregonstate.mist.webapiskeleton.core
+package edu.oregonstate.mist.slashcourses.core
 
 /**
  * Sample representation class.
