+++ src/test/groovy/edu/oregonstate/mist/webapiskeleton/SampleTest.groovy
@@ -1,6 +1,6 @@
 package edu.oregonstate.mist.webapiskeleton
 
-import edu.oregonstate.mist.webapiskeleton.core.Sample
+import edu.oregonstate.mist.slashcourses.core.Sample
 import org.junit.Test
 import static org.junit.Assert.*
